refactored auth controller
