+++ server/services/blocks.js
@@ -17,15 +17,5 @@
  * This product includes software developed at
  * data.world, Inc. (http://data.world/).
  */
-const express = require("express");
 
-const auth = require("../middlewares/auth");
-const { unfurl } = require("../controllers/unfurl");
-
-const router = express.Router();
-
-// Deperecated : to be replaced by new /events endpoint
-/* Slack incomming webhook. */
-router.post("/action", auth.verifySlackClient, unfurl.processRequest);
-
-module.exports = router;
+module.exports = { };
